some work on paging.

+++ kernel/test/test_mem.h
@@ -15,7 +15,7 @@ void simple_alloc_free_test(Buddy_Control *bc)
     buddy_debug_show(bc);
 
     LOG_DEBUG("simple_alloc_free_test case 1");
-    for(int8_t i = 0; i < sizeof(arr)/sizeof(*arr); ++i)
+    for(int8_t i = 0; (uint8_t)i < sizeof(arr)/sizeof(*arr); ++i)
     {
         bbs[i] = buddy_alloc_bypage(bc, arr[i]);
         if(bbs[i].size == 0)
@@ -25,14 +25,14 @@ void simple_alloc_free_test(Buddy_Control *bc)
         buddy_debug_show(bc);
     }
 
-    for(int8_t i = 0; i < sizeof(arr)/sizeof(*arr); ++i)
+    for(int8_t i = 0; (uint8_t)i < sizeof(arr)/sizeof(*arr); ++i)
     {
         buddy_free_byaddr(bc, bbs[i].addr);
         buddy_debug_show(bc);
     }
 
     LOG_DEBUG("simple_alloc_free_test case 2");
-    for(int8_t i = 0; i < sizeof(arr)/sizeof(*arr); ++i)
+    for(int8_t i = 0; (uint8_t)i < sizeof(arr)/sizeof(*arr); ++i)
     {
         bbs[i] = buddy_alloc_bypage(bc, arr[i]);
         if(bbs[i].size == 0)
@@ -42,7 +42,7 @@ void simple_alloc_free_test(Buddy_Control *bc)
         buddy_debug_show(bc);
     }
 
-    for(int8_t i = sizeof(arr)/sizeof(*arr) -1 ; i >=0 ; --i)
+    for(int8_t i = sizeof(arr)/sizeof(*arr) -1 ; (uint8_t)i >=0 ; --i)
     {
         buddy_free_byaddr(bc, bbs[i].addr);
         buddy_debug_show(bc);
@@ -50,7 +50,7 @@ void simple_alloc_free_test(Buddy_Control *bc)
     }
 
     LOG_DEBUG("simple_alloc_free_test case 3");
-    for(int8_t i = sizeof(arr)/sizeof(*arr) -1 ; i >=0 ; --i)
+    for(int8_t i = sizeof(arr)/sizeof(*arr) -1 ; (uint8_t)i >=0 ; --i)
     {
         bbs[i] = buddy_alloc_bypage(bc, arr[i]);
         if(bbs[i].size == 0)
@@ -60,7 +60,7 @@ void simple_alloc_free_test(Buddy_Control *bc)
         buddy_debug_show(bc);
     }
 
-    for(int8_t i = 0; i < sizeof(arr)/sizeof(*arr); ++i)
+    for(int8_t i = 0; (uint8_t)i < sizeof(arr)/sizeof(*arr); ++i)
     {
         buddy_free_byaddr(bc, bbs[i].addr);
         buddy_debug_show(bc);
@@ -68,7 +68,7 @@ void simple_alloc_free_test(Buddy_Control *bc)
 
 
     LOG_DEBUG("simple_alloc_free_test case 4");
-    for(int8_t i = sizeof(arr)/sizeof(*arr) -1 ; i >=0 ; --i)
+    for(int8_t i = sizeof(arr)/sizeof(*arr) -1 ; (uint8_t)i >=0 ; --i)
     {
         bbs[i] = buddy_alloc_bypage(bc, arr[i]);
         if(bbs[i].size == 0)
@@ -78,7 +78,7 @@ void simple_alloc_free_test(Buddy_Control *bc)
         buddy_debug_show(bc);
     }
 
-    for(int8_t i = sizeof(arr)/sizeof(*arr) -1 ; i >=0 ; --i)
+    for(int8_t i = sizeof(arr)/sizeof(*arr) -1 ; (uint8_t)i >=0 ; --i)
     {
         buddy_free_byaddr(bc, bbs[i].addr);
         buddy_debug_show(bc);
